requirements.txt
httpx
beautifulsoup4
PyGithub
orjson
requests
